Added some things
Moved original main.rs to _main.rs
Added Debug to Data struct
Made a mostly functional prototype with axum
