+++ Cargo.toml
@@ -10,7 +10,7 @@ repository = "https://github.com/4198-Programmers/meanscout-api-rust"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-rocket = {version = "0.5.0-rc", features = ["tls", "json"]}
+# rocket = {version = "0.5.0-rc", features = ["tls", "json"]}
 # rocket_contrib = { version = "0.4.11", features = ["tera_templates"]}
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
@@ -20,6 +20,5 @@ config = { version = "0.13.3" }
 tera = "1.18.1"
 lazy_static = "1.4.0"
 csv = { version = "1.2.1" }
-
-[patch.crates-io]
-traitobject = { git = "https://github.com/reem/rust-traitobject", rev = "b3471a15917b2caf5a8b27debb0b4b390fc6634f" }
+axum = { git = "https://github.com/tokio-rs/axum.git", features = ["json"] }
+tower-http = { version = "0.4.0", features = ["cors"] }
